+++ custom_addons/subscription_suite/subscription_suite_billing/tests/__init__.py
@@ -5,5 +5,6 @@ from . import test_manager_operations
 from . import test_deferred_revenue
 from . import test_security_access
 from . import test_performance_smoke
+from . import test_operation_observability
 from . import test_performance_smoke
 from . import test_security_access