+++ custom_addons/subscription_suite/subscription_suite_dunning/models/__init__.py
@@ -1,6 +1,7 @@
 from . import dunning_policy
 from . import dunning_policy_line
 from . import subscription_dunning_attempt
+from . import subscription_operation_run
 from . import subscription_manager_operation_dashboard
 from . import subscription_plan
 from . import sale_order