+++ versions/19.0/custom_addons/subscription_suite_billing/models/__init__.py
@@ -1,5 +1,6 @@
 from . import subscription_proration
 from . import subscription_billing_run
+from . import subscription_operation_run
 from . import subscription_billing_attempt
 from . import subscription_payment_attempt
 from . import subscription_plan_change_request